+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.25.0] - 2021-10-08
+
 ### Fixed
 
 - `report` subcommands were showing only the time not the date when the time entry was created.
@@ -687,7 +689,8 @@ time entry.
 - Golang CLI using [cobra](https://github.com/spf13/cobra)
 - Makefile to help setup actions
 
-[Unreleased]: https://github.com/lucassabreu/clockify-cli/compare/v0.24.1...HEAD
+[Unreleased]: https://github.com/lucassabreu/clockify-cli/compare/v0.25.0...HEAD
+[v0.24.1]: https://github.com/lucassabreu/clockify-cli/releases/tag/v0.24.1
 [v0.24.0]: https://github.com/lucassabreu/clockify-cli/releases/tag/v0.24.0
 [v0.23.1]: https://github.com/lucassabreu/clockify-cli/releases/tag/v0.23.1
 [v0.23.0]: https://github.com/lucassabreu/clockify-cli/releases/tag/v0.23.0